Automatically scale navigation grid to vessel size
The grid is still a cube centered on the vessel's root part, which can leave a
lot of unused space. It may be worth experimenting with further adjusting grid
bounds by making the integer grid rectangular. There is still also a noticable
delay in pathfinding near the edge of the navigation grid, so we may need to
further tweak default grid size.